pyproject.toml
[tool.poetry]
name = "doc-reader"
version = "0.1.0"
description = ""
authors = ["遥奕 <jibing.yjb@alibaba-inc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pdfplumber = "^0.11.0"
pyspellchecker = "^0.8.1"
pymupdf = "^1.24.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
ocrmypdf = "^16.1.2"
nltk = "^3.8.1"
ftfy = "^6.2.0"
python-magic = "^0.4.27"
pyfunvice = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
